+++ src/frameworks/opengles/util.rs
@@ -177,3 +177,71 @@ pub fn try_decode_pvrtc(
     };
     true
 }
+
+pub struct PalettedTextureFormat {
+    /// `true` for 4-bit (nibble) index, 16-color palette. `false` for 8-bit (byte) index, 256-color palette.
+    pub index_is_nibble: bool,
+    /// `glTexImage2D`-style `format` for palette entries: `GL_RGB` or `GL_RGBA`
+    pub palette_entry_format: GLenum,
+    /// `glTexImage2D`-style `type` for palette entries: `GL_UNSIGNED_BYTE` or some `GL_UNSIGNED_SHORT_` value
+    pub palette_entry_type: GLenum,
+}
+impl PalettedTextureFormat {
+    /// If the provided format is from `OES_compressed_paletted_texture`, returns
+    /// [Some] with information about it, or [None] otherwise.
+    pub fn get_info(internalformat: GLenum) -> Option<Self> {
+        match internalformat {
+            gles11::PALETTE4_RGB8_OES => Some(Self {
+                index_is_nibble: true,
+                palette_entry_format: gles11::RGB,
+                palette_entry_type: gles11::UNSIGNED_BYTE,
+            }),
+            gles11::PALETTE4_RGBA8_OES => Some(Self {
+                index_is_nibble: true,
+                palette_entry_format: gles11::RGBA,
+                palette_entry_type: gles11::UNSIGNED_BYTE,
+            }),
+            gles11::PALETTE4_R5_G6_B5_OES => Some(Self {
+                index_is_nibble: true,
+                palette_entry_format: gles11::RGB,
+                palette_entry_type: gles11::UNSIGNED_SHORT_5_6_5,
+            }),
+            gles11::PALETTE4_RGBA4_OES => Some(Self {
+                index_is_nibble: true,
+                palette_entry_format: gles11::RGBA,
+                palette_entry_type: gles11::UNSIGNED_SHORT_4_4_4_4,
+            }),
+            gles11::PALETTE4_RGB5_A1_OES => Some(Self {
+                index_is_nibble: true,
+                palette_entry_format: gles11::RGBA,
+                palette_entry_type: gles11::UNSIGNED_SHORT_5_5_5_1,
+            }),
+            gles11::PALETTE8_RGB8_OES => Some(Self {
+                index_is_nibble: false,
+                palette_entry_format: gles11::RGB,
+                palette_entry_type: gles11::UNSIGNED_BYTE,
+            }),
+            gles11::PALETTE8_RGBA8_OES => Some(Self {
+                index_is_nibble: false,
+                palette_entry_format: gles11::RGBA,
+                palette_entry_type: gles11::UNSIGNED_BYTE,
+            }),
+            gles11::PALETTE8_R5_G6_B5_OES => Some(Self {
+                index_is_nibble: false,
+                palette_entry_format: gles11::RGB,
+                palette_entry_type: gles11::UNSIGNED_SHORT_5_6_5,
+            }),
+            gles11::PALETTE8_RGBA4_OES => Some(Self {
+                index_is_nibble: false,
+                palette_entry_format: gles11::RGBA,
+                palette_entry_type: gles11::UNSIGNED_SHORT_4_4_4_4,
+            }),
+            gles11::PALETTE8_RGB5_A1_OES => Some(Self {
+                index_is_nibble: false,
+                palette_entry_format: gles11::RGBA,
+                palette_entry_type: gles11::UNSIGNED_SHORT_5_5_5_1,
+            }),
+            _ => None,
+        }
+    }
+}