+++ src/window/gl_bindings/build.rs
@@ -42,6 +42,8 @@ fn main() {
             "GL_EXT_texture_filter_anisotropic",
             "GL_IMG_texture_compression_pvrtc",
             "GL_EXT_texture_lod_bias",
+            // Part of the OpenGL ES 1.1 common profile.
+            "GL_OES_compressed_paletted_texture",
         ],
     )
     .write_bindings(GlobalGenerator, &mut file)